feat(email): centralize campaign type
