+++ functions/marketing-email-sender.ts
@@ -1,6 +1,7 @@
-import { sendDueCampaigns } from "@acme/email";
+import { sendDueCampaigns, type Campaign } from "@acme/email";
 
 export { sendDueCampaigns };
+export type { Campaign };
 
 export const onScheduled = async () => {
   await sendDueCampaigns();
