+++ packages/email/src/storage/types.ts
@@ -1,13 +1,4 @@
-export interface Campaign {
-  id: string;
-  recipients: string[];
-  subject: string;
-  body: string;
-  segment?: string | null;
-  sendAt: string;
-  sentAt?: string;
-  templateId?: string | null;
-}
+import type { Campaign } from "../types";
 
 export interface CampaignStore {
   /**
