+++ packages/email/src/types.ts
@@ -1,0 +1,10 @@
+export interface Campaign {
+  id: string;
+  recipients: string[];
+  subject: string;
+  body: string;
+  segment?: string | null;
+  sendAt: string;
+  sentAt?: string;
+  templateId?: string | null;
+}
